+++ applications/vessel-history/src/features/map/map.selectors.ts
@@ -1,32 +1,37 @@
 import { createSelector } from '@reduxjs/toolkit'
-import { Generators } from '@globalfishingwatch/layer-composer'
-import { BACKGROUND_LAYER, DEFAULT_DATAVIEWS } from 'data/config'
-import { selectHiddenLayers, selectSatellite } from 'routes/routes.selectors'
+// import { Generators } from '@globalfishingwatch/layer-composer'
+import { getDataviewsGeneratorConfigs } from '@globalfishingwatch/dataviews-client'
+// import { selectHiddenLayers, selectSatellite } from 'routes/routes.selectors'
+import { selectDataviewInstancesResolved } from 'features/dataviews/dataviews.selectors'
+import { BACKGROUND_LAYER, OFFLINE_LAYERS } from 'features/dataviews/dataviews.config'
+
+const generatorOptions = {
+  // heatmapAnimatedMode,
+  // highlightedTime,
+  // timeRange: staticTime,
+  // debug: debugOptions.debug,
+  // mergedActivityGeneratorId: MERGED_ACTIVITY_ANIMATED_HEATMAP_GENERATOR_ID,
+}
+
+// const generatorsConfig = getDataviewsGeneratorConfigs(dataviews, generatorOptions, resources)
 
 /**
  * Select the base layers that are not hidden by the user
  */
 export const selectMapLayers = createSelector(
-  [selectHiddenLayers, selectSatellite],
-  (hiddenLayers, satellite) => {
-    const dataviews: any = DEFAULT_DATAVIEWS.map((dataview) => {
-      return {
-        ...dataview,
-        basemap:
-          dataview.type !== Generators.Type.Basemap
-            ? dataview.type
-            : satellite
-            ? Generators.BasemapType.Satellite
-            : Generators.BasemapType.Default,
-        visible: !hiddenLayers.includes(dataview.id),
-      }
-    })
-    return dataviews
+  [selectDataviewInstancesResolved],
+  (dataviewsInstances) => {
+    if (!dataviewsInstances) return
+    return getDataviewsGeneratorConfigs(
+      dataviewsInstances,
+      generatorOptions
+      // resources
+    )
   }
 )
 /**
  * Merge all the layers needed to render the map
  */
-export const getLayerComposerLayers = createSelector([selectMapLayers], (mapLayers) => {
-  return [...BACKGROUND_LAYER, ...mapLayers]
+export const getLayerComposerLayers = createSelector([selectMapLayers], (mapLayers = []) => {
+  return [...BACKGROUND_LAYER, ...OFFLINE_LAYERS, ...mapLayers]
 })