+++ packages/dataviews-client/src/resolve-dataviews-generators.ts
@@ -130,6 +130,7 @@ export function getGeneratorConfig(
           {
             id: generator.id,
             colorRamp: dataview.config?.colorRamp as ColorRampsIds,
+            colorRampWhiteEnd: false,
             visible: dataview.config?.visible ?? true,
             breaks: dataview.config?.breaks,
             datasets: datasetsIds,
@@ -264,6 +265,7 @@ export function getDataviewsGeneratorConfigs(
         id: dataview.id,
         datasets,
         colorRamp: config.colorRamp as Generators.ColorRampsIds,
+        colorRampWhiteEnd: true,
         filter: config.filter,
         visible: config.visible,
         legend: {