+++ applications/fishing-map/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @globalfishingwatchapp/fishing-map
 
+## 0.28.34
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+- Updated dependencies [undefined]
+  - @globalfishingwatch/ui-components@2.18.12
+  - @globalfishingwatch/api-types@3.3.0
+
+## 0.28.33
+
+### Patch Changes
+
+- e94c28f7: Fishingmap/feedback popup
+- Updated dependencies [e94c28f7]
+  - @globalfishingwatch/ui-components@2.18.11
+
+## 0.28.32
+
+### Patch Changes
+
+- Updated dependencies [cd532944]
+- Updated dependencies [cd532944]
+  - @globalfishingwatch/dataviews-client@6.2.4
+  - @globalfishingwatch/layer-composer@5.2.5
+  - @globalfishingwatch/ui-components@2.18.10
+  - @globalfishingwatch/react-hooks@6.0.10
+
 ## 0.28.31
 
 ### Patch Changes