+++ packages/ui-components/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @globalfishingwatch/ui-components
 
+## 2.18.12
+
+### Patch Changes
+
+- update marine manager icon
+
+## 2.18.11
+
+### Patch Changes
+
+- e94c28f7: Fishingmap/feedback popup
+
+## 2.18.10
+
+### Patch Changes
+
+- cd532944: fix when no icon found
+- Updated dependencies [cd532944]
+  - @globalfishingwatch/layer-composer@5.2.5
+
 ## 2.18.9
 
 ### Patch Changes